Patched clock.ts's dependancies and merged localclockstuff back into it
